Fixes from testing towerside code on hardware

+++ src/towerside/sensors.cpp
@@ -6,11 +6,11 @@
 namespace sensors {
 
 uint16_t get_main_batt_mv() {
-  return analogRead(pinout::MAIN_BATT_VOLTAGE) * 48 / 17 * 5;
+  return static_cast<uint16_t>(analogRead(pinout::MAIN_BATT_VOLTAGE)) * 48 / 17 * 5;
 }
 
 uint16_t get_actuator_batt_mv() {
-  return analogRead(pinout::ACTUATOR_BATT_VOLTAGE) * 48 / 17 * 5;
+  return static_cast<uint16_t>(analogRead(pinout::ACTUATOR_BATT_VOLTAGE)) * 48 / 17 * 5;
 }
 
 uint16_t get_towerside_state() {
